size_t are unsigned ints. Fixes #244

+++ src/parse.cpp
@@ -13,11 +13,12 @@ IntegerVector dim_tokens_(List sourceSpec, List tokenizerSpec) {
   TokenizerPtr tokenizer = Tokenizer::create(tokenizerSpec);
   tokenizer->tokenize(source->begin(), source->end());
 
-  size_t rows = -1, cols = -1;
+  int rows = -1, cols = -1;
 
   for (Token t = tokenizer->nextToken(); t.type() != TOKEN_EOF; t = tokenizer->nextToken()) {
     rows = t.row();
-    if (t.col() > cols)
+
+    if ((int) t.col() > cols)
       cols = t.col();
   }
